The resulting executable still needs to be tested against 32 bit and 64 bit Sender.

+++ APSICommon/apsi/ffield/ffield_array.h
@@ -28,7 +28,7 @@ namespace apsi
             field_(field)
         {
             // Initialize array
-            array_.resize(field_.d_ * size_, 0);
+            array_.resize(static_cast<size_t>(field_.d_ * size_), 0);
         }
 
         FFieldArray(const FFieldArray &copy) = default;
